Extract simple AST checks to src/ast/checks.ts
